orchestration/pyproject.toml
[project]
name = "crank-orchestration"
version = "0.1.0"
description = "LangGraph orchestrator for the Crank scaffold. Drives layer subgraphs and recurses over them with explicit loop edges."
requires-python = ">=3.11"

dependencies = [
  "langgraph==0.4.0",
  "langchain-anthropic==0.3.5",
  "claude-agent-sdk==0.5.0",
  "pydantic==2.9.2",
  "pyyaml==6.0.2",
  "jsonschema==4.23.0",
  "click==8.1.7",
  "rich==13.9.4",
]

[project.optional-dependencies]
dev = [
  "pytest==8.3.3",
  "pytest-asyncio==0.24.0",
  "ruff==0.7.0",
]

[project.scripts]
crank-run = "orchestration.run:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["orchestration"]
